SQL 대기 계산의 overflow를 막고 붙임 조건의 축약을 정본에서 없앤다
- SQL 대기 계산이 bigint 로 먼저 곱해 상한(LEAST)에 닿기 전에 overflow 가 날 수 있었다. 기본값
  에서는 문제없지만 설정은 큰 기준 대기와 높은 상한을 허용하고 문서도 그렇게 적었다. numeric 으로
  계산해 상한을 적용한 뒤 변환한다. 도메인 구현은 상한에서 곱셈 전에 반환해 이미 안전하다.
- 03·계약13·PLAN 이 붙임 조건을 "단일 후보 + 제출 원장 대응"으로 축약했다. 그대로 재구현하면
  같은 금액의 다른 자산 이동을 붙일 수 있다. 관찰 일치(지갑·자산 키·금액·목적지)를 함께 적고,
  목적지 비교가 EVM 에서만 대소문자를 무시한다는 것도 적는다.
- 발신을 "미판단 계열·후속"으로 적은 계약13 세 곳과 워커 주석을 현재 동작으로 맞춘다.
- WebhookRetryBackoff KDoc 이 형식 변환을 호출자가 한다고 했으나 지금은 저장소 SQL 이 한다.
  이 객체가 식의 정본이고 두 구현을 영속 테스트가 묶는다고 적는다.

+++ blockchain-manager-infra/persistence/src/main/kotlin/com/whatto/bcm/infra/persistence/webhook/WebhookInboxJdbcAdapter.kt
@@ -128,10 +128,12 @@ class WebhookInboxJdbcAdapter(
                                ELSE to_char(
                                    to_timestamp(:now, 'YYYYMMDDHH24MISS')
                                        + make_interval(
+                                           -- numeric으로 계산해 **상한을 적용하기 전에** overflow가 나지 않게 한다.
+                                           -- bigint로 먼저 곱하면 큰 기준 대기·높은 상한에서 LEAST에 닿기 전에 넘친다.
                                            secs => LEAST(
-                                               :maxSeconds::bigint,
-                                               :baseSeconds::bigint * power(2, rtry_cnt)::bigint
-                                           )
+                                               :maxSeconds::numeric,
+                                               :baseSeconds::numeric * power(2::numeric, rtry_cnt)
+                                           )::bigint
                                        ),
                                    'YYYYMMDDHH24MISS'
                                )
